requirements.txt
aiofiles==23.1.0
aioredis==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
flatten-dict==0.2.0
hiredis==2.2.2
html5tagger==1.3.0
httptools==0.5.0
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==5.2.0
password==0.2
pathlib2==2.3.7.post1
PyJWT==2.7.0
pypika-tortoise==0.1.6
pytz==2023.3
PyYAML==6.0
sanic==23.3.0
sanic-routing==22.8.0
six==1.16.0
tortoise-orm==0.19.3
tracerite==1.1.0
typing_extensions==4.5.0
ujson==5.7.0
uvloop==0.17.0
websockets==10.4
